renaming files to fit style guide
