+++ experimental/reporting-ui/src/main/react/reporting-ui/library/reportingClient/models/list_reports_response.ts
@@ -12,14 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import '@testing-library/jest-dom';
-import '@testing-library/jest-dom/extend-expect';
-import { render, screen } from '@testing-library/react';
-import React from 'react';
-import App from '../../../main/react/reporting-ui/src/App';
+import {Report} from './report';
 
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
+export type ListReportsResponse = {
+  reports: Report[];
+};
