+++ experimental/reporting-ui/src/main/react/reporting-ui/library/reportingClient/models/summary_publisher_data.ts
