+++ experimental/reporting-ui/src/main/react/reporting-ui/src/index.tsx
@@ -15,8 +15,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import App from './app';
+import reportWebVitals from './report_web_vitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root')!);
 root.render(
