+++ experimental/reporting-ui/src/test/react/reporting-ui/app.test.tsx
@@ -12,8 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {Report} from './Report';
+import '@testing-library/jest-dom';
+import '@testing-library/jest-dom/extend-expect';
+import { render, screen } from '@testing-library/react';
+import React from 'react';
+import App from '../../../main/react/reporting-ui/src/app';
 
-export type GetReportResponse = {
-  report?: Report;
-};
+test('renders learn react link', () => {
+  render(<App />);
+  const linkElement = screen.getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
+});
